% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateGenes.hg20.R
\name{annotateGenes.hg20}
\alias{annotateGenes.hg20}
\title{annotate genes with reference to hg38.}
\usage{
annotateGenes.hg20(mat, annotation, ID.type = "S")
}
\arguments{
\item{mat}{data matrix; genes in rows; cell names in columns.}

\item{annotation}{annotation file for all known genes.}

\item{ID.type}{gene id type: Symbol or Ensemble.}
}
\value{
annotations of each genes in rows with chrom and positions.
}
\description{
annotate genes with reference to hg38.
}
\examples{
test.anno.mat <- annotateGenes.hg20(mat=matx, ID.type="ENSEMBLE_id", full.anno = full.anno)
}
