% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateGenes.mm10.R
\name{annotateGenes.mm10}
\alias{annotateGenes.mm10}
\title{annotate genes with reference to hg38.}
\usage{
annotateGenes.mm10(mat, ID.type = "S", full.anno = full.anno.mm10)
}
\arguments{
\item{mat}{data matrix; genes in rows; cell names in columns.}

\item{ID.type}{gene id type: Symbol or Ensemble.}

\item{full.anno}{annotation file for all known genes, automatically loaded in copycat.}
}
\value{
annotations of each genes in rows with chrom and positions.
}
\description{
annotate genes with reference to hg38.
}
\examples{
test.anno.mat <- annotateGenes.hg20(mat=matx, ID.type="ENSEMBLE_id", full.anno = full.anno)
}
