% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateGenes.R
\name{annotateGenes}
\alias{annotateGenes}
\title{annotate genes with reference to hg38.}
\usage{
annotateGenes(mat, annotation, ID.type = "S")
}
\arguments{
\item{mat}{data matrix; genes in rows; cell names in columns.}

\item{annotation}{annotation file for all known genes.}

\item{ID.type}{gene id type: Symbol or Ensemble.}
}
\value{
annotations of each genes in rows with chrom and positions.
}
\description{
annotate genes with reference to hg38.
}
\examples{
test.anno.mat <- annotateGenes.hg20(mat=matx,annotation=annotation, ID.type="ENSEMBLE_id")
}
