% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.norm.cl.R
\name{baseline.norm.cl}
\alias{baseline.norm.cl}
\title{find a cluster of diploid cells with integrative clustering method}
\usage{
baseline.norm.cl(norm.mat.smooth, min.cells = 5, n.cores = n.cores)
}
\arguments{
\item{norm.mat.smooth}{smoothed data matrix; genes in rows; cell names in columns.}

\item{min.cells}{minimal number of cells per cluster.}

\item{n.cores}{number of cores for parallel computing.}
}
\value{
1) predefined diploid cell names; 2) clustering results; 3) inferred baseline.
}
\description{
find a cluster of diploid cells with integrative clustering method
}
\examples{
test.bnc <- baseline.norm.cl(norm.mat.smooth=norm.mat.smooth, min.cells=5, n.cores=10)

test.bnc.cells <- test.bnc$preN
}
