% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.synthetic.R
\name{baseline.synthetic}
\alias{baseline.synthetic}
\title{estimate basline copy numbers by synthetic normal cell using GMM to estimate sigmas}
\usage{
baseline.synthetic(norm.mat = norm.mat, min.cells = 10, n.cores)
}
\arguments{
\item{norm.mat}{smoothed data matrix; genes in rows; cell names in columns.}

\item{min.cells}{minimal number of cells per cluster.}

\item{n.cores}{number of cores for parallel computing.}
}
\value{
1) relative gene expression; 2) synthetic baseline profiles; 3) clustering results.
}
\description{
estimate basline copy numbers by synthetic normal cell using GMM to estimate sigmas
}
\examples{
test.relt <- baseline.synthetic(norm.mat=orm.mat.smooth, min.cells=10, n.cores)

norm.mat.relat <- test.relt$expr.relat
}
