% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_distance.R
\name{cal_dist}
\alias{cal_dist}
\title{Calculate Parallel Distance Matrix via Fork-Based mclapply}
\usage{
cal_dist(mat, method = "emd", num_cores = NULL)
}
\arguments{
\item{mat}{A matrix where rows are genomic bins and columns are cells.}

\item{method}{Character string: "emd", "euclidean", "pearson", or "spearman".}

\item{num_cores}{Integer specifying CPU cores. Defaults to total available cores minus 1.}
}
\value{
A formal R 'dist' class object.
}
\description{
Calculate Parallel Distance Matrix via Fork-Based mclapply
}
