% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNA.MCMC.R
\name{CNA.MCMC}
\alias{CNA.MCMC}
\title{MCMC segmentation.}
\usage{
CNA.MCMC(clu, fttmat, bins, cut.cor, n.cores)
}
\arguments{
\item{clu}{predefined clusters of cells.}

\item{fttmat}{transformed normalized data matrix; genes in rows; cell names in columns.}

\item{bins}{minimal number of bins per segment.}

\item{cut.cor}{minimal KS difference to call breakpoints.}

\item{n.cores}{number of cores for parallel computing.}

\item{sam.name}{given sample name.}
}
\value{
1) segmented CNA data matrix; 2) vector of breakpoints.
}
\description{
MCMC segmentation.
}
\examples{
test.mcmc <- CNA.MCMC(clu,fttmat, bins, cut.cor, n.cores=10)

test.mcmc.seg.mat <- test.mamc$logCNA
}
