% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.all.bins.hg20.R
\name{convert.all.bins.hg20}
\alias{convert.all.bins.hg20}
\title{conver gene by cell matrix to genomic bins by cells matrix.}
\usage{
convert.all.bins.hg20(DNA.mat, RNA.mat, n.cores)
}
\arguments{
\item{DNA.mat}{input target bins, provided by copycat, with 220KB windows.}

\item{RNA.mat}{RNA data matrix with genes in rows.}

\item{n.cores}{number of cores for parallel computing.}

\item{full.anno}{annotation file for all known genes, automatically loaded in copycat.}
}
\value{
adjusted datamatrix with bins in rows, cells in columns.
}
\description{
conver gene by cell matrix to genomic bins by cells matrix.
}
\examples{
test.cab <- convert.all.bins(DNA.mat, RNA.mat, n.cores=10)

test.cab.uber <- test.cab$RNAadj
}
