% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copykat.R
\name{copykat}
\alias{copykat}
\title{copycat main_func.}
\usage{
copykat(
  rawmat = rawdata,
  id.type = "S",
  ngene.chr = 0,
  LOW.DR = 0.1,
  UP.DR = 0.2,
  win.size = 25,
  norm.cell.names = "",
  KS.cut = 0.15,
  sam.name = "",
  n.cores = 1
)
}
\arguments{
\item{rawmat}{raw data matrix; genes in rows; cell names in columns.}

\item{id.type}{gene id type: Symbol or Ensemble.}

\item{ngene.chr}{minimal number of genes per chromosome for cell filtering.}

\item{LOW.DR}{minimal population fractoins of genes for smoothing.}

\item{UP.DR}{minimal population fractoins of genes for segmentation.}

\item{win.size}{minimal window sizes for segmentation.}

\item{norm.cell.names}{a vector of normal cell names.}

\item{KS.cut}{segmentation parameters, input 0 to 1; larger looser criteria.}

\item{sam.name}{sample name.}

\item{n.cores}{number of cores for parallel computing.}
}
\value{
1) aneuploid/diploid prediction results; 2) CNA results in 220KB windows; 3) heatmap; 4) hclustering object.
}
\description{
copycat main_func.
}
\examples{
test.ck <- copykat(rawmat=rawdata, sam.name="test", n.cores=10)

test.pred <- test.ck$prediction

}
