% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp.rawdata.R
\docType{data}
\name{exp.rawdata}
\alias{exp.rawdata}
\title{raw UMI count matrix.}
\format{
A data matrix with 33694 rows (genes) and 1097 columns (single cell names):
}
\usage{
exp.rawdata
}
\description{
An example dataset containing the UMI counts for 300 single cells.
}
\keyword{datasets}
